requirements.txt
Django==5.0.6
djangorestframework==3.15.1
drf-yasg==1.21.7
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
flower==2.0.1
